* Fixed up several scripts with Qwen, which also allowed me delete some unneeded ones
